Added domain entities

+++ ContentManagement.Entities/UserRole.cs
@@ -6,9 +6,8 @@ namespace ContentManagement.Entities
 {
     public class UserRole : IEntityType
     {
-        public int UserId { get; set; }
-        public int RoleId { get; set; }
-
+        public long UserId       { get; set; }
+        public int RoleId        { get; set; }
         public virtual User User { get; set; }
         public virtual Role Role { get; set; }
     }
