+++ ContentManagement.Services/Contracts/IRolesService.cs
@@ -8,8 +8,8 @@ namespace ContentManagement.Services.Contracts
 {
     public interface IRolesService
     {
-        Task<List<Role>> FindUserRolesAsync(int userId);
-        Task<bool> IsUserInRole(int userId, string roleName);
+        Task<List<Role>> FindUserRolesAsync(long userId);
+        Task<bool> IsUserInRole(long userId, string roleName);
         Task<List<User>> FindUsersInRoleAsync(string roleName);
     }
 }
